refactor(filters): move filter declarations to `filters` package
