+++ filters/gaussian_blur/action.go
@@ -7,9 +7,9 @@
 // (represented by sigma, or "σ").
 package gaussian_blur
 
-import "dougdomingos.com/image-filters/filters"
+import "dougdomingos.com/image-filters/filters/types"
 
-var GaussianBlurAction = filters.Action{
+var GaussianBlurAction = types.Action{
 	Preprocess: nil,
 	Filter:     GaussianBlur,
 }
