+++ filters/horizontal_flip/action.go
@@ -4,11 +4,9 @@
 // effectively mirroring the image along its vertical axis.
 package horizontal_flip
 
-import (
-	"dougdomingos.com/image-filters/filters"
-)
+import "dougdomingos.com/image-filters/filters/types"
 
-var HorizontalFlipAction = filters.Action{
+var HorizontalFlipAction = types.Action{
 	Preprocess: nil,
 	Filter:     HorizontalFlip,
 }
