+++ pipelines/types.go
@@ -4,11 +4,12 @@ import (
 	"fmt"
 
 	"dougdomingos.com/image-filters/filters"
+	"dougdomingos.com/image-filters/filters/types"
 )
 
 type Step struct {
-	Action *filters.Action
-	Next *Step
+	Action *types.Action
+	Next   *Step
 }
 
 type Recipe struct {
@@ -29,7 +30,7 @@ func BuildRecipe(filterIDs []string) (Recipe, error) {
 	var head, tail *Step
 
 	for _, id := range filterIDs {
-		pipeline, err := GetFilterPipeline(id)
+		pipeline, err := filters.GetFilterAction(id)
 		if err != nil {
 			return Recipe{}, fmt.Errorf("[ERROR] No pipeline found for filter \"%q\"", id)
 		}
